Creada clase base e impĺementados los metodos que eran comunes entre los recursos
